requirements.txt
# Core MCP dependencies
mcp>=1.0.0
# Note: MCP requires pydantic>=2.11.0,<3.0.0
pydantic>=2.11.0,<3.0.0
# Note: MCP requires httpx>=0.27.1
httpx>=0.27.1

# Google API dependencies
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
# Google API client with all service extras
google-api-python-client[gmail,drive,calendar,sheets,docs,photos,contacts]>=2.100.0
google-cloud-core>=2.3.0

# Data processing and export
pandas>=2.0.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# Email processing
email-validator>=2.0.0
mail-parser>=4.0.0

# Scheduling and automation
croniter>=1.4.0
schedule>=1.2.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
toml>=0.10.0

# Logging and monitoring
structlog>=23.1.0
colorama>=0.4.6

# Encryption and security
cryptography>=41.0.0
keyring>=24.0.0

# Async support
# Note: asyncio is built into Python 3.4+, no package needed
aiofiles>=23.1.0
aiohttp>=3.8.0

# CLI and user interface
click>=8.1.0
rich>=13.0.0
typer>=0.9.0
